Refactors CompleteMoabHandler.
refs #1964
