+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config --auto-gen-only-exclude`
-# on 2022-10-20 12:10:43 UTC using RuboCop version 1.36.0.
+# on 2022-10-21 20:17:00 UTC using RuboCop version 1.36.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -14,14 +14,15 @@ Layout/LineLength:
   Exclude:
     - 'app/models/druid_version_zip.rb'
 
-# Offense count: 36
+# Offense count: 32
 # Configuration parameters: AllowedMethods, AllowedPatterns, IgnoredMethods.
 Lint/AmbiguousBlockAssociation:
   Exclude:
     - 'spec/jobs/moab_replication_audit_job_spec.rb'
-    - 'spec/services/complete_moab_handler_check_exist_spec.rb'
-    - 'spec/services/complete_moab_handler_confirm_version_spec.rb'
-    - 'spec/services/complete_moab_handler_update_version_spec.rb'
+    - 'spec/services/complete_moab_service/check_existence_spec.rb'
+    - 'spec/services/complete_moab_service/shared_examples.rb'
+    - 'spec/services/complete_moab_service/update_version_after_validation_spec.rb'
+    - 'spec/services/complete_moab_service/update_version_spec.rb'
     - 'spec/services/shared_examples_complete_moab_handler.rb'
 
 # Offense count: 2
@@ -30,54 +31,44 @@ Lint/MissingSuper:
     - 'app/services/reporters/logger_reporter.rb'
     - 'config/initializers/okcomputer.rb'
 
-# Offense count: 25
-# Configuration parameters: AllowedMethods, AllowedPatterns, IgnoredMethods, CountRepeatedAttributes, Max.
+# Offense count: 23
+# Configuration parameters: AllowedMethods, AllowedPatterns, IgnoredMethods, CountRepeatedAttributes.
 Metrics/AbcSize:
-  Exclude:
-    - 'app/controllers/catalog_controller.rb'
-    - 'app/controllers/objects_controller.rb'
-    - 'app/services/audit/catalog_to_archive.rb'
-    - 'app/services/audit/catalog_to_moab.rb'
-    - 'app/services/audit/checksum.rb'
-    - 'app/services/audit/moab_to_catalog.rb'
-    - 'app/services/s3/s3_audit.rb'
-    - 'app/models/druid_version_zip.rb'
-    - 'app/models/preserved_object.rb'
-    - 'app/services/checksum_validator.rb'
-    - 'app/services/complete_moab_handler.rb'
-    - 'app/services/moab_storage_root_reporter.rb'
+  Max: 66
 
 # Offense count: 2
 # Configuration parameters: CountBlocks, Max.
 Metrics/BlockNesting:
   Exclude:
-    - 'app/services/complete_moab_handler.rb'
+    - 'app/services/complete_moab_service/update_version_after_validation.rb'
 
-# Offense count: 4
+# Offense count: 2
 # Configuration parameters: CountComments, Max, CountAsOne.
 Metrics/ClassLength:
   Exclude:
     - 'app/controllers/objects_controller.rb'
     - 'app/models/druid_version_zip.rb'
-    - 'app/services/checksum_validator.rb'
-    - 'app/services/complete_moab_handler.rb'
 
 # Offense count: 3
 # Configuration parameters: AllowedMethods, AllowedPatterns, IgnoredMethods, Max.
 Metrics/CyclomaticComplexity:
   Exclude:
-    - 'app/services/complete_moab_handler.rb'
+    - 'app/services/complete_moab_service/check_existence.rb'
+    - 'app/services/complete_moab_service/update_version.rb'
+    - 'app/services/complete_moab_service/update_version_after_validation.rb'
 
-# Offense count: 41
+# Offense count: 40
 # Configuration parameters: CountComments, CountAsOne, ExcludedMethods, AllowedMethods, AllowedPatterns, IgnoredMethods.
 Metrics/MethodLength:
-  Max: 32
+  Max: 33
 
 # Offense count: 3
 # Configuration parameters: AllowedMethods, AllowedPatterns, IgnoredMethods, Max.
 Metrics/PerceivedComplexity:
   Exclude:
-    - 'app/services/complete_moab_handler.rb'
+    - 'app/services/complete_moab_service/check_existence.rb'
+    - 'app/services/complete_moab_service/update_version.rb'
+    - 'app/services/complete_moab_service/update_version_after_validation.rb'
 
 # Offense count: 3
 # Configuration parameters: EnforcedStyle, CheckMethodNames, CheckSymbols, AllowedIdentifiers, AllowedPatterns.
@@ -98,17 +89,12 @@ RSpec/InstanceVariable:
   Exclude:
     - 'spec/models/moab_storage_directory_spec.rb'
 
-# Offense count: 2
-RSpec/LetSetup:
-  Exclude:
-    - 'spec/services/complete_moab_handler_confirm_version_spec.rb'
-
-# Offense count: 407
+# Offense count: 378
 # Configuration parameters: AllowSubject.
 RSpec/MultipleMemoizedHelpers:
   Max: 21
 
-# Offense count: 18
+# Offense count: 15
 RSpec/StubbedMock:
   Exclude:
     - 'spec/jobs/catalog_to_moab_job_spec.rb'
@@ -116,13 +102,11 @@ RSpec/StubbedMock:
     - 'spec/jobs/moab_to_catalog_job_spec.rb'
     - 'spec/jobs/part_replication_audit_job_spec.rb'
     - 'spec/services/audit/catalog_to_moab_spec.rb'
-    - 'spec/services/audit/moab_to_catalog_spec.rb'
-    - 'spec/services/s3/shared_examples_s3_audit.rb'
     - 'spec/services/checksum_validator_spec.rb'
-    - 'spec/services/complete_moab_handler_check_exist_spec.rb'
-    - 'spec/services/complete_moab_handler_create_spec.rb'
-    - 'spec/services/complete_moab_handler_spec.rb'
-    - 'spec/services/complete_moab_handler_update_version_spec.rb'
+    - 'spec/services/complete_moab_service/check_existence_spec.rb'
+    - 'spec/services/complete_moab_service/create_after_validation_spec.rb'
+    - 'spec/services/complete_moab_service/update_version_after_validation_spec.rb'
+    - 'spec/services/s3/shared_examples_s3_audit.rb'
 
 # Offense count: 4
 RSpec/SubjectDeclaration:
@@ -180,8 +164,8 @@ Rails/TimeZone:
 # Offense count: 6
 Style/CombinableLoops:
   Exclude:
-    - 'spec/services/audit/catalog_to_moab_spec.rb'
     - 'spec/models/moab_storage_directory_spec.rb'
+    - 'spec/services/audit/catalog_to_moab_spec.rb'
 
 # Offense count: 5
 # Configuration parameters: AllowedConstants.
@@ -193,18 +177,11 @@ Style/Documentation:
     - 'config/application.rb'
 
 # Offense count: 1
-# Configuration parameters: MinBodyLength, AllowConsecutiveConditionals.
-Style/GuardClause:
-  Exclude:
-    - 'app/services/complete_moab_handler.rb'
-
-# Offense count: 8
 # Configuration parameters: AllowedMethods.
 # AllowedMethods: respond_to_missing?
 Style/OptionalBooleanParameter:
   Exclude:
     - 'app/services/s3/s3_audit.rb'
-    - 'app/services/complete_moab_handler.rb'
 
 # Offense count: 4
 # This cop supports unsafe autocorrection (--autocorrect-all).
