+++ spec/services/complete_moab_service/create_after_validation_spec.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 require 'rails_helper'
-require 'services/shared_examples_complete_moab_handler'
+require 'services/complete_moab_service/shared_examples'
 
-RSpec.describe CompleteMoabHandler do
+RSpec.describe CompleteMoabService::CreateAfterValidation do
   let(:druid) { 'ab123cd4567' }
   let(:incoming_version) { 6 }
   let(:incoming_size) { 9876 }
@@ -23,89 +23,23 @@ RSpec.describe CompleteMoabHandler do
     allow(Reporters::EventServiceReporter).to receive(:new).and_return(event_service_reporter)
   end
 
-  describe '#create' do
-    it 'creates PreservedObject and CompleteMoab and PreservedObjectsPrimaryMoab in database' do
-      complete_moab_handler.create
-      new_po = PreservedObject.find_by(druid: druid)
-      new_cm = new_po.complete_moab
-      expect(new_po.current_version).to eq incoming_version
-      expect(new_cm.moab_storage_root).to eq ms_root
-      expect(new_cm.size).to eq incoming_size
-      expect(new_po.preserved_objects_primary_moab.complete_moab_id).to eq new_cm.id
-    end
-
-    it 'creates the CompleteMoab with "ok" status and validation timestamps if caller ran CV' do
-      complete_moab_handler.create(true)
-      new_cm = complete_moab_handler.pres_object.complete_moab
-      expect(new_cm.status).to eq 'ok'
-      expect(new_cm.last_version_audit).to be_a ActiveSupport::TimeWithZone
-      expect(new_cm.last_moab_validation).to be_a ActiveSupport::TimeWithZone
-      expect(new_cm.last_checksum_validation).to be_a ActiveSupport::TimeWithZone
-    end
-
-    it_behaves_like 'attributes validated', :create
-
-    it 'object already exists' do
-      complete_moab_handler.create
-      new_complete_moab_handler = described_class.new(druid, incoming_version, incoming_size, ms_root)
-      results = new_complete_moab_handler.create
-      code = AuditResults::DB_OBJ_ALREADY_EXISTS
-      expect(results).to include(a_hash_including(code => a_string_matching('CompleteMoab db object already exists')))
-    end
-
-    it_behaves_like 'calls AuditResults.report_results', :create
-
-    context 'db update error' do
-      context 'ActiveRecordError' do
-        before do
-          allow(PreservedObject).to receive(:create!).with(hash_including(druid: druid))
-                                                     .and_raise(ActiveRecord::ActiveRecordError, 'foo')
-        end
-
-        it 'DB_UPDATE_FAILED result' do
-          expect(complete_moab_handler.create).to include(a_hash_including(AuditResults::DB_UPDATE_FAILED))
-        end
-
-        it 'does NOT get CREATED_NEW_OBJECT result' do
-          expect(complete_moab_handler.create).not_to include(hash_including(AuditResults::CREATED_NEW_OBJECT))
-        end
-      end
-
-      it "rolls back PreservedObject creation if the CompleteMoab can't be created (e.g. due to DB constraint violation)" do
-        po = instance_double(PreservedObject)
-        allow(PreservedObject).to receive(:create!).with(hash_including(druid: druid)).and_return(po)
-        allow(po).to receive(:create_complete_moab!).and_raise(ActiveRecord::RecordInvalid)
-        complete_moab_handler.create
-        expect(PreservedObject.find_by(druid: druid)).to be_nil
-      end
-    end
-
-    context 'returns' do
-      let(:result) { complete_moab_handler.create }
-
-      it '1 result of CREATED_NEW_OBJECT' do
-        expect(result).to be_an_instance_of Array
-        expect(result.size).to eq 1
-        expect(result.first).to match(a_hash_including(AuditResults::CREATED_NEW_OBJECT => exp_msg))
-      end
-    end
-  end
-
-  describe '#create_after_validation' do
+  describe '#execute' do
     let(:valid_druid) { 'bp628nk4868' }
     let(:storage_dir) { 'spec/fixtures/storage_root02/sdr2objects' }
-    let(:complete_moab_handler) { described_class.new(valid_druid, incoming_version, incoming_size, ms_root) }
+    let(:complete_moab_handler) do
+      described_class.new(druid: valid_druid, incoming_version: incoming_version, incoming_size: incoming_size, moab_storage_root: ms_root)
+    end
 
-    it_behaves_like 'attributes validated', :create_after_validation
+    it_behaves_like 'attributes validated'
 
-    it_behaves_like 'calls AuditResults.report_results', :create_after_validation
+    it_behaves_like 'calls AuditResultsReporter.report_results'
 
     context 'sets validation timestamps' do
       let(:t) { Time.current }
       let(:ms_root) { MoabStorageRoot.find_by(storage_location: storage_dir) }
       let(:cm_db_obj) { complete_moab_handler.pres_object.complete_moab }
 
-      before { complete_moab_handler.create_after_validation }
+      before { complete_moab_handler.execute }
 
       it 'sets last_moab_validation with current time' do
         expect(cm_db_obj.last_moab_validation).to be_within(10).of(t)
@@ -117,8 +51,9 @@ RSpec.describe CompleteMoabHandler do
     end
 
     it 'creates PreservedObject and CompleteMoab and PreservedObjectsPrimaryMoab in database when there are no validation errors' do
-      complete_moab_handler = described_class.new(valid_druid, incoming_version, incoming_size, ms_root)
-      complete_moab_handler.create_after_validation
+      complete_moab_handler = described_class.new(druid: valid_druid, incoming_version: incoming_version, incoming_size: incoming_size,
+                                                  moab_storage_root: ms_root)
+      complete_moab_handler.execute
       new_po = PreservedObject.find_by(druid: valid_druid, current_version: incoming_version)
       expect(new_po).not_to be_nil
       new_cm = new_po.complete_moab
@@ -128,8 +63,9 @@ RSpec.describe CompleteMoabHandler do
     end
 
     it 'creates CompleteMoab with "ok" status and validation timestamps if no validation errors and caller ran CV' do
-      complete_moab_handler = described_class.new(valid_druid, incoming_version, incoming_size, ms_root)
-      complete_moab_handler.create_after_validation(true)
+      complete_moab_handler = described_class.new(druid: valid_druid, incoming_version: incoming_version, incoming_size: incoming_size,
+                                                  moab_storage_root: ms_root)
+      complete_moab_handler.execute(checksums_validated: true)
       new_po = PreservedObject.find_by(druid: valid_druid, current_version: incoming_version)
       expect(new_po).not_to be_nil
       new_cm = new_po.complete_moab
@@ -142,15 +78,18 @@ RSpec.describe CompleteMoabHandler do
       mock_sov = instance_double(Stanford::StorageObjectValidator)
       expect(mock_sov).to receive(:validation_errors).and_return([])
       allow(Stanford::StorageObjectValidator).to receive(:new).and_return(mock_sov)
-      complete_moab_handler = described_class.new(valid_druid, incoming_version, incoming_size, ms_root)
-      complete_moab_handler.create_after_validation
+      complete_moab_handler = described_class.new(druid: valid_druid, incoming_version: incoming_version, incoming_size: incoming_size,
+                                                  moab_storage_root: ms_root)
+      complete_moab_handler.execute
     end
 
     context 'when moab is invalid' do
       let(:storage_dir) { 'spec/fixtures/bad_root01/bad_moab_storage_trunk' }
       let(:ms_root) { MoabStorageRoot.find_by(storage_location: storage_dir) }
       let(:invalid_druid) { 'xx000xx0000' }
-      let(:complete_moab_handler) { described_class.new(invalid_druid, incoming_version, incoming_size, ms_root) }
+      let(:complete_moab_handler) do
+        described_class.new(druid: invalid_druid, incoming_version: incoming_version, incoming_size: incoming_size, moab_storage_root: ms_root)
+      end
 
       # add storage root with invalid moab to the MoabStorageRoots table
       before do
@@ -160,7 +99,7 @@ RSpec.describe CompleteMoabHandler do
       end
 
       it 'creates PreservedObject, and CompleteMoab with "invalid_moab" status, and PreservedObjectsPrimaryMoab in database' do
-        complete_moab_handler.create_after_validation
+        complete_moab_handler.execute
         new_po = PreservedObject.find_by(druid: invalid_druid, current_version: incoming_version)
         expect(new_po).not_to be_nil
         new_cm = new_po.complete_moab
@@ -172,7 +111,7 @@ RSpec.describe CompleteMoabHandler do
       end
 
       it 'creates CompleteMoab with "invalid_moab" status in database even if caller ran CV' do
-        complete_moab_handler.create_after_validation(true)
+        complete_moab_handler.execute(checksums_validated: true)
         new_po = PreservedObject.find_by(druid: invalid_druid, current_version: incoming_version)
         expect(new_po).not_to be_nil
         new_cm = new_po.complete_moab
@@ -181,7 +120,7 @@ RSpec.describe CompleteMoabHandler do
       end
 
       it 'includes invalid moab result' do
-        results = complete_moab_handler.create_after_validation
+        results = complete_moab_handler.execute.results
         expect(results).to include(a_hash_including(AuditResults::INVALID_MOAB => /Invalid Moab, validation errors:/))
       end
 
@@ -190,8 +129,9 @@ RSpec.describe CompleteMoabHandler do
           let(:results) do
             allow(PreservedObject).to receive(:create!).with(hash_including(druid: invalid_druid))
                                                        .and_raise(ActiveRecord::ActiveRecordError, 'foo')
-            complete_moab_handler = described_class.new(invalid_druid, incoming_version, incoming_size, ms_root)
-            complete_moab_handler.create_after_validation
+            complete_moab_handler = described_class.new(druid: invalid_druid, incoming_version: incoming_version, incoming_size: incoming_size,
+                                                        moab_storage_root: ms_root)
+            complete_moab_handler.execute.results
           end
 
           it 'DB_UPDATE_FAILED result' do
@@ -205,20 +145,22 @@ RSpec.describe CompleteMoabHandler do
 
         it "rolls back PreservedObject creation if the CompleteMoab can't be created (e.g. due to DB constraint violation)" do
           allow(CompleteMoab).to receive(:create!).and_raise(ActiveRecord::RecordInvalid)
-          complete_moab_handler = described_class.new(invalid_druid, incoming_version, incoming_size, ms_root)
-          complete_moab_handler.create
+          complete_moab_handler = described_class.new(druid: invalid_druid, incoming_version: incoming_version, incoming_size: incoming_size,
+                                                      moab_storage_root: ms_root)
+          complete_moab_handler.execute
           expect(PreservedObject.where(druid: druid)).not_to exist
         end
       end
     end
 
     context 'returns' do
-      let(:result) { complete_moab_handler.create_after_validation }
+      let(:audit_result) { complete_moab_handler.execute }
+      let(:results) { audit_result.results }
 
       it '1 CREATED_NEW_OBJECT result' do
-        expect(result).to be_an_instance_of Array
-        expect(result.size).to eq 1
-        expect(result.first).to include(AuditResults::CREATED_NEW_OBJECT => exp_msg)
+        expect(audit_result).to be_an_instance_of AuditResults
+        expect(results.size).to eq 1
+        expect(results.first).to include(AuditResults::CREATED_NEW_OBJECT => exp_msg)
       end
     end
   end
